Refactor store collector
