+++ silica/visitors/__init__.py
@@ -1,1 +1,2 @@
-from silica.visitors.collect_names import collect_names
+from .collect_names import collect_names
+from .collect_stores import collect_stores
